+++ src/fn_call.rs
@@ -321,7 +321,7 @@ impl Engine {
             return Ok(match name {
                 KEYWORD_PRINT => {
                     if let Some(ref print) = self.print {
-                        let text = result.as_immutable_string().map_err(|typ| {
+                        let text = result.into_immutable_string().map_err(|typ| {
                             EvalAltResult::ErrorMismatchOutputType(
                                 self.map_type_name(type_name::<ImmutableString>()).into(),
                                 typ.into(),
@@ -335,7 +335,7 @@ impl Engine {
                 }
                 KEYWORD_DEBUG => {
                     if let Some(ref debug) = self.debug {
-                        let text = result.as_immutable_string().map_err(|typ| {
+                        let text = result.into_immutable_string().map_err(|typ| {
                             EvalAltResult::ErrorMismatchOutputType(
                                 self.map_type_name(type_name::<ImmutableString>()).into(),
                                 typ.into(),
@@ -1100,7 +1100,7 @@ impl Engine {
 
                 // Fn - only in function call style
                 return arg
-                    .as_immutable_string()
+                    .into_immutable_string()
                     .map_err(|typ| self.make_type_mismatch_err::<ImmutableString>(typ, arg_pos))
                     .and_then(FnPtr::try_from)
                     .map(Into::<Dynamic>::into)
@@ -1150,7 +1150,7 @@ impl Engine {
                 )?;
 
                 let fn_name = arg
-                    .as_immutable_string()
+                    .into_immutable_string()
                     .map_err(|typ| self.make_type_mismatch_err::<ImmutableString>(typ, arg_pos))?;
 
                 let (arg, arg_pos) = self.get_arg_value(
@@ -1176,7 +1176,7 @@ impl Engine {
                     scope, mods, state, lib, this_ptr, level, args_expr, constants, 0,
                 )?;
                 let var_name = arg
-                    .as_immutable_string()
+                    .into_immutable_string()
                     .map_err(|typ| self.make_type_mismatch_err::<ImmutableString>(typ, arg_pos))?;
                 return Ok(scope.contains(&var_name).into());
             }
@@ -1189,7 +1189,7 @@ impl Engine {
                     scope, mods, state, lib, this_ptr, level, args_expr, constants, 0,
                 )?;
                 let script = &value
-                    .as_immutable_string()
+                    .into_immutable_string()
                     .map_err(|typ| self.make_type_mismatch_err::<ImmutableString>(typ, pos))?;
                 let result =
                     self.eval_script_expr_in_place(scope, mods, state, lib, script, pos, level + 1);