+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Enhancements
 ------------
 
 * `$symbol$` is supported in custom syntax to match any symbol.
+* `Dynamic::as_string` and `Dynamic::as_immutable_string` are deprecated and replaced by `into_string` and `into_immutable_string` respectively.
 * `parse_float()`, `PI()` and `E()` now defer to `Decimal` under `no_float` if `decimal` is turned on.